services/backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
fastapi==0.124.0
h11==0.16.0
idna==3.11
psycopg==3.3.2
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
bcrypt==4.1.2

# Add these missing dependencies
python-jose[cryptography]==3.3.0
passlib==1.7.4
sqlalchemy==2.0.25
python-multipart==0.0.6
pydantic-settings==2.1.0
cryptography==41.0.7